Rename fixtures package to mock
