+++ mgmt/rest/v2/mock/mock_config_manager.go
@@ -18,7 +18,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package fixtures
+package mock
 
 import (
 	"github.com/intelsdi-x/snap/core"
